RFPredict/requirements.txt
# Function dependencies, for example:
# package>=version
requests==2.24.0
scikit-learn>=0.24.2
google-cloud-storage==1.25.0
numpy
keras==2.14.0
#tensorflow==2.14.0
pandas
